requirements.txt
openai
requests
tinydb
pillow